+++ modulo3/cookenu/public/manifest.json
