+++ modulo3/cookenu/src/App.js
@@ -1,0 +1,12 @@
+import React from "react"
+import {Router} from "react-router-dom"
+
+const App = () => {
+  return (
+    <div>
+     <Router/>     
+    </div>
+  )  
+}
+
+export default App
