+++ modulo3/cookenu/src/pages/AdicionarReceita/AdicionarReceita.js
@@ -3,12 +3,10 @@ import React from "react";
 const AdicionarReceita = () => {
     return (
         <div>
-            <h1>
-                AdicionarReceita
-            </h1>
+            <h1>AdicionarReceita</h1>
         </div>
     )
 
-}
+};
 
 export default AdicionarReceita
