+++ modulo3/cookenu/src/pages/Cadastro/Cadastro.js
@@ -6,7 +6,6 @@ const Cadastro = () => {
             <h1>Cadastro</h1>
         </div>
     )
-
 }
 
 export default Cadastro
