+++ modulo3/cookenu/src/pages/DetalheReceita/DetalheReceita.js
@@ -3,12 +3,10 @@ import React from "react";
 const DetalheReceita = () => {
     return (
         <div>
-            <h1>
-                DetalheReceita
-            </h1>
+            <h1>DetalheReceita </h1>
         </div>
     )
 
-}
+};
 
 export default DetalheReceita
