+++ modulo3/cookenu/src/pages/ListaReceita/ListaReceita.js
@@ -3,11 +3,9 @@ import React from "react";
 const ListaReceita = () => {
     return (
         <div>
-            <h1>
-                ListaReceita
-            </h1>
+            <h1>ListaReceita</h1>       
         </div>
     )
 
-}
+};
 export default ListaReceita
