+++ modulo3/cookenu/src/pages/Login/Login.js
@@ -3,12 +3,10 @@ import React from "react";
 const Login = () => {
     return (
         <div>
-            <h1>
-                Login
-            </h1>
+            <h1>Login</h1>    
         </div>
     )
 
-}
+};
 
 export default Login
