+++ modulo3/cookenu/src/pages/PaginaErro/PaginaErro.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const PaginaErro = () => {
+    return (
+        <div>
+            <h1> PaginaErro</h1>        
+        </div>
+    )
+
+}
+
+export default PaginaErro
